Update issue 77
Request/Response services based on a SOAP --> JMS pattern can now be generated including sample transformers between XML and a CSV format.
Unit-tests classes are generated for the transformers and integration test is generated for the full service including negative tests for invalid input and timeouts.
Based on the sample-wsdl in soi-toolkits refapps.